Refactoring some things
Attempting to extend the config system to allow for default values, type-checking, and valid ranges. I've chosen to combine this with some refactoring to slim down common.py to make it more manageable.
