+++ dice.py
